requirements.txt
# Pinned to avoid RequestsDependencyWarning noise.
# (Requests warns when it detects an unsupported urllib3/chardet/charset_normalizer combo.)
requests==2.32.5
urllib3==2.6.3
charset-normalizer==3.4.4
chardet==5.2.0

readability-lxml==0.8.4.1
html2text==2025.4.15
playwright==1.58.0
